+++ sdk/communication/Azure.Communication.PhoneNumbers/src/PhoneNumbersClient.cs
@@ -105,7 +105,7 @@ namespace Azure.Communication.PhoneNumbers
         }
         #endregion protected constructors
 
-        /// <summary> Releases an acquired phone number. </summary>
+        /// <summary> Releases an purchased phone number. </summary>
         /// <param name="phoneNumber"> Phone number to be released, e.g. +14255550123. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <exception cref="ArgumentNullException"> <paramref name="phoneNumber"/> is null. </exception>
@@ -127,7 +127,7 @@ namespace Azure.Communication.PhoneNumbers
             }
         }
 
-        /// <summary> Releases an acquired phone number. </summary>
+        /// <summary> Releases an purchased phone number. </summary>
         /// <param name="phoneNumber"> Phone number to be released, e.g. +14255550123. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <exception cref="ArgumentNullException"> <paramref name="phoneNumber"/> is null. </exception>
@@ -197,12 +197,12 @@ namespace Azure.Communication.PhoneNumbers
             }
         }
 
-        /// <summary> Gets the details of the given acquired phone number. </summary>
+        /// <summary> Gets the details of the given purchased phone number. </summary>
         /// <param name="phoneNumber"> The acquired phone number whose details are to be fetched in E.164 format, e.g. +14255550123. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
-        public virtual async Task<Response<AcquiredPhoneNumber>> GetPhoneNumberAsync(string phoneNumber, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<PurchasedPhoneNumber>> GetPurchasedPhoneNumberAsync(string phoneNumber, CancellationToken cancellationToken = default)
         {
-            using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPhoneNumber)}");
+            using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPurchasedPhoneNumber)}");
             scope.Start();
             try
             {
@@ -215,12 +215,12 @@ namespace Azure.Communication.PhoneNumbers
             }
         }
 
-        /// <summary> Gets the details of the given acquired phone number. </summary>
+        /// <summary> Gets the details of the given purchased phone number. </summary>
         /// <param name="phoneNumber"> The acquired phone number whose details are to be fetched in E.164 format, e.g. +14255550123. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
-        public virtual Response<AcquiredPhoneNumber> GetPhoneNumber(string phoneNumber, CancellationToken cancellationToken = default)
+        public virtual Response<PurchasedPhoneNumber> GetPurchasedPhoneNumber(string phoneNumber, CancellationToken cancellationToken = default)
         {
-            using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPhoneNumber)}");
+            using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPurchasedPhoneNumber)}");
             scope.Start();
             try
             {
@@ -329,13 +329,13 @@ namespace Azure.Communication.PhoneNumbers
             }
         }
 
-        /// <summary> Gets the list of all acquired phone numbers. </summary>
+        /// <summary> Gets the list of all purchased phone numbers. </summary>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
-        public virtual AsyncPageable<AcquiredPhoneNumber> GetPhoneNumbersAsync(CancellationToken cancellationToken = default)
+        public virtual AsyncPageable<PurchasedPhoneNumber> GetPurchasedPhoneNumbersAsync(CancellationToken cancellationToken = default)
         {
-            async Task<Page<AcquiredPhoneNumber>> FirstPageFunc(int? pageSizeHint)
+            async Task<Page<PurchasedPhoneNumber>> FirstPageFunc(int? pageSizeHint)
             {
-                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPhoneNumbers)}");
+                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPurchasedPhoneNumbers)}");
                 scope.Start();
                 try
                 {
@@ -348,9 +348,9 @@ namespace Azure.Communication.PhoneNumbers
                     throw;
                 }
             }
-            async Task<Page<AcquiredPhoneNumber>> NextPageFunc(string nextLink, int? pageSizeHint)
+            async Task<Page<PurchasedPhoneNumber>> NextPageFunc(string nextLink, int? pageSizeHint)
             {
-                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPhoneNumbers)}");
+                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPurchasedPhoneNumbers)}");
                 scope.Start();
                 try
                 {
@@ -366,13 +366,13 @@ namespace Azure.Communication.PhoneNumbers
             return PageableHelpers.CreateAsyncEnumerable(FirstPageFunc, NextPageFunc);
         }
 
-        /// <summary> Gets the list of all acquired phone numbers. </summary>
+        /// <summary> Gets the list of all purchased phone numbers. </summary>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
-        public virtual Pageable<AcquiredPhoneNumber> GetPhoneNumbers(CancellationToken cancellationToken = default)
+        public virtual Pageable<PurchasedPhoneNumber> GetPurchasedPhoneNumbers(CancellationToken cancellationToken = default)
         {
-            Page<AcquiredPhoneNumber> FirstPageFunc(int? pageSizeHint)
+            Page<PurchasedPhoneNumber> FirstPageFunc(int? pageSizeHint)
             {
-                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPhoneNumbers)}");
+                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPurchasedPhoneNumbers)}");
                 scope.Start();
                 try
                 {
@@ -385,9 +385,9 @@ namespace Azure.Communication.PhoneNumbers
                     throw;
                 }
             }
-            Page<AcquiredPhoneNumber> NextPageFunc(string nextLink, int? pageSizeHint)
+            Page<PurchasedPhoneNumber> NextPageFunc(string nextLink, int? pageSizeHint)
             {
-                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPhoneNumbers)}");
+                using var scope = _clientDiagnostics.CreateScope($"{nameof(PhoneNumbersClient)}.{nameof(GetPurchasedPhoneNumbers)}");
                 scope.Start();
                 try
                 {